requirements.txt
altgraph==0.17.4
numpy==2.2.2
packaging==24.2
pandas==2.2.3
pefile==2023.2.7
pillow==11.1.0
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
python-dateutil==2.9.0.post0
pytz==2025.1
pywin32-ctypes==0.2.3
six==1.17.0
tk==0.1.0
tzdata==2025.1
